backend/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.6.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cryptography==36.0.0
cycler==0.11.0
distlib==0.3.5
Django==4.0.6
djangorestframework==3.13.1
django-heroku==0.3.1
django-mptt==0.13.4
django-environ==0.9.0
filelock==3.8.0
fonttools==4.33.3
gunicorn==20.1.0
h11==0.12.0
idna==3.2
Jinja2==3.0.3
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pathlib2==2.3.6
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
terminaltables==3.1.0
toml==0.10.2
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.6
user-agent==0.1.10
virtualenv==20.16.3
wsproto==1.0.0
django-allauth==0.51.0
django-cors-headers==3.13.0
dj-rest-auth==2.2.5
djangorestframework-simplejwt==5.2.0 
psycopg2==2.9.3
isort==5.10.1
django-js-asset==2.0.0
drf-spectacular>=0.15.1,<0.16